+++ moirai-iter/src/parallel.rs
@@ -21,8 +21,8 @@ mod traits;
 
 pub use adapters::{
     Chain, Chunks, Cloned, Copied, Enumerate, Filter, FilterMap, FlatMap, Flatten, Inspect,
-    Intersperse, Map, MapInit, MapPositions, MapWith, PanicFuse, Positions, Rev, Skip,
-    SkipAnyWhile, Take, TakeAnyWhile, Update, WhileSome, Zip, ZipEq,
+    Interleave, InterleaveShortest, Intersperse, Map, MapInit, MapPositions, MapWith, PanicFuse,
+    Positions, Rev, Skip, SkipAnyWhile, Take, TakeAnyWhile, Update, WhileSome, Zip, ZipEq,
 };
 pub use consumers::Reduction;
 pub use fallible::TryStreamItem;